pyproject.toml
[project]
name = "arc-diffusion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flax>=0.12.0",
    "grain>=0.2.13",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "optax>=0.2.6",
    "sws-config>=0.6.0",
    "tqdm>=4.67.1",
    "wandb>=0.22.3",
]

[project.optional-dependencies]
cpu = ["jax>=0.7.0"]
gpu = ["jax[cuda12]>=0.7.0"]
tpu = ["jax[tpu]>=0.7.0"]
